requirement.txt
Flask==2.0.3
Flask-MySQLdb==1.0.1
mysqlclient==2.1.1
numpy==1.23.1
opencv-python==4.6.0.66
python-dotenv==0.20.0
requests==2.28.1
Werkzeug==2.0.3